+++ src/brandedVoice/index.tsx
@@ -1,0 +1,1 @@
+export { BrandedVoice, BrandedVoiceProps } from './BrandedVoice'