+++ src/newItinerary/HiddenStop.style.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+import { color, space } from '../_utils/branding'
+import { TextCaption } from '../typography/caption'
+
+export const StyledLabel = styled(TextCaption)`
+  padding: ${space.m} 0;
+  color: ${color.midnightGreen};
+`