+++ src/successModal/SuccessModal.style.tsx
@@ -3,6 +3,7 @@ import styled from 'styled-components'
 import { color, componentSizes, responsiveBreakpoints, space } from '../_utils/branding'
 import { Button } from '../button'
 import { Modal } from '../modal'
+import { TheVoice } from '../theVoice'
 
 const StyledSuccessModal = styled(Modal)`
   padding: 0;
@@ -81,16 +82,17 @@ const Content = styled.div`
 
   @media (${responsiveBreakpoints.isMediaLarge}) {
     height: auto;
+    /* gutter */
+    margin-left: ${space.l};
+    max-width: ${componentSizes.smallSectionWidth};
   }
 `
 
 /* SuccessModal Spacements */
-const SuccessTitle = styled.h1`
+const SuccessTitle = styled(TheVoice).attrs({ isInverted: true })`
   display: flex;
   flex: 1;
   justify-content: center;
-  margin: 0; /*reset h1*/
-  padding: ${space.xl} 0 32px; /* TheVoice spacing*/
   white-space: pre-line;
 `